+++ entityx/Event.cc
@@ -21,7 +21,7 @@ EventManager::~EventManager() {
 }
 
 void EventManager::emit(const BaseEvent &event) {
-  auto sig = signal_for(event.my_family());
+  auto sig = signal_for(static_cast<int>(event.my_family()));
   sig->emit(&event);
 }
